95% working OS X PLTLabs API RC1.

+++ Desktop/OS X/PLTDevice/PLTDevice/BRPedometerSettingResponse.h
@@ -11,6 +11,6 @@
 
 @interface BRPedometerSettingResponse : BRSettingResponse
 
-@property(nonatomic,readonly) uint16_t steps;
+@property(nonatomic,readonly) uint32_t steps;
 
 @end
